+++ artemis-core/src/main/java/org/jnosql/artemis/column/query/ColumnQueryDeleteParser.java
@@ -25,7 +25,7 @@ import static org.jnosql.artemis.column.query.ColumnQueryParserUtil.toCondition;
 
 /**
  * Class the returns a {@link ColumnDeleteQuery}
- * on {@link ColumnCrudRepositoryProxy}
+ * on {@link ColumnRepositoryProxy}
  */
 public class ColumnQueryDeleteParser {
